Change deprecated methods.

+++ NGMoviePlayer/NGMoviePlayer/NGScrubber.m
@@ -203,7 +203,7 @@
         CGFloat trackingOffset = currentLocation.x - previousLocation.x;
         
         // Find the scrubbing speed that curresponds to the touch's vertical offset
-        CGFloat verticalOffset = fabsf(currentLocation.y - self.beganTrackingLocation.y);
+        CGFloat verticalOffset = fabs(currentLocation.y - self.beganTrackingLocation.y);
         NSUInteger scrubbingSpeedChangePosIndex = [self indexOfLowerScrubbingSpeed:self.scrubbingSpeedChangePositions forOffset:verticalOffset];        
         
         if (scrubbingSpeedChangePosIndex == NSNotFound) {
@@ -221,7 +221,7 @@
         if (((self.beganTrackingLocation.y < currentLocation.y) && (currentLocation.y < previousLocation.y)) ||
             ((self.beganTrackingLocation.y > currentLocation.y) && (currentLocation.y > previousLocation.y)) ) {
             // We are getting closer to the slider, go closer to the real location
-			thumbAdjustment = (_realPositionValue - self.value) / ( 1 + fabsf(currentLocation.y - self.beganTrackingLocation.y));
+			thumbAdjustment = (_realPositionValue - self.value) / ( 1 + fabs(currentLocation.y - self.beganTrackingLocation.y));
         }
         
 		self.value += valueAdjustment + thumbAdjustment;
